+++ src/main/redux/sagas/reader.ts
@@ -106,6 +106,7 @@ function* closeReader(reader: Reader, gotoLibrary: boolean) {
 
     if (typedAction.error) {
         // Failed to close publication
+        // FIXME : this action is linked to nothing
         yield put(readerActions.closeError.build(reader));
         return;
     }
