+++ src/main/streamer.ts
@@ -9,13 +9,11 @@ import * as debug_ from "debug";
 import { app } from "electron";
 import * as express from "express";
 import * as path from "path";
+import * as portfinder from "portfinder";
+import { computeReadiumCssJsonMessage } from "readium-desktop/common/computeReadiumCssJsonMessage";
 import { diMainGet } from "readium-desktop/main/di";
 import { _NODE_MODULE_RELATIVE_URL, _PACKAGING } from "readium-desktop/preprocessor-directives";
 
-import { IEventPayload_R2_EVENT_READIUMCSS } from "@r2-navigator-js/electron/common/events";
-import {
-    colCountEnum, IReadiumCSS, readiumCSSDefaults, textAlignEnum,
-} from "@r2-navigator-js/electron/common/readium-css-settings";
 import { setupReadiumCSS } from "@r2-navigator-js/electron/main/readium-css";
 import { secureSessions } from "@r2-navigator-js/electron/main/sessions";
 import { Publication as R2Publication } from "@r2-shared-js/models/publication";
@@ -26,133 +24,70 @@ import { Server } from "@r2-streamer-js/http/server";
 
 const debug = debug_("readium-desktop:main#streamer");
 
-// Create readium2 streamer
-// This streamer is used to stream epub content to the renderer
-export const streamer: Server = new Server({
-    disableDecryption: false,
-    disableOPDS: true,
-    disableReaders: true,
-    disableRemotePubUrl: true,
-});
+function streamerFactory(): Server {
 
-app.on("ready", () => {
-    secureSessions(streamer); // HTTPS
-});
+    // Create readium2 streamer
+    // This streamer is used to stream epub content to the renderer
+    const streamer: Server = new Server({
+        disableDecryption: false,
+        disableOPDS: true,
+        disableReaders: true,
+        disableRemotePubUrl: true,
+    });
 
-let rcssPath = "ReadiumCSS";
-if (_PACKAGING === "1") {
-    rcssPath = path.normalize(path.join(__dirname, rcssPath));
-} else {
-    rcssPath = "r2-navigator-js/dist/ReadiumCSS";
-    rcssPath = path.normalize(path.join(__dirname, _NODE_MODULE_RELATIVE_URL, rcssPath));
-}
+    app.on("ready", () => {
+        secureSessions(streamer); // HTTPS
 
-rcssPath = rcssPath.replace(/\\/g, "/");
-debug("readium css path:", rcssPath);
+        let rcssPath = "ReadiumCSS";
+        if (_PACKAGING === "1") {
+            rcssPath = path.normalize(path.join(__dirname, rcssPath));
+        } else {
+            rcssPath = "r2-navigator-js/dist/ReadiumCSS";
+            rcssPath = path.normalize(path.join(__dirname, _NODE_MODULE_RELATIVE_URL, rcssPath));
+        }
 
-// TODO: centralize this code, currently duplicated
-// see src/renderer/components/reader/ReaderApp.jsx
-function computeReadiumCssJsonMessage(_r2Publication: R2Publication, _link: Link | undefined):
-    IEventPayload_R2_EVENT_READIUMCSS {
-    const store = diMainGet("store");
-    const settings = store.getState().reader.config;
-    debug(settings);
+        rcssPath = rcssPath.replace(/\\/g, "/");
+        debug("readium css path:", rcssPath);
 
-    // TODO: see the readiumCSSDefaults values below, replace with readium-desktop's own
-    const cssJson: IReadiumCSS = {
+        const store = diMainGet("store");
+        const setting = store.getState().reader.config;
+        setupReadiumCSS(streamer, rcssPath, computeReadiumCssJsonMessage(setting));
+    });
 
-        a11yNormalize: readiumCSSDefaults.a11yNormalize,
-
-        backgroundColor: readiumCSSDefaults.backgroundColor,
-
-        bodyHyphens: readiumCSSDefaults.bodyHyphens,
-
-        colCount: settings.colCount === "1" ? colCountEnum.one :
-            (settings.colCount === "2" ? colCountEnum.two : colCountEnum.auto),
-
-        darken: settings.dark,
-
-        font: settings.font,
-
-        fontSize: settings.fontSize,
-
-        invert: settings.invert,
-
-        letterSpacing: settings.letterSpacing,
-
-        ligatures: readiumCSSDefaults.ligatures,
-
-        lineHeight: settings.lineHeight,
-
-        night: settings.night,
-
-        pageMargins: settings.pageMargins,
-
-        paged: settings.paged,
-
-        paraIndent: readiumCSSDefaults.paraIndent,
-
-        paraSpacing: settings.paraSpacing,
-
-        sepia: settings.sepia,
-
-        noFootnotes: settings.noFootnotes,
-
-        textAlign: settings.align === textAlignEnum.left ? textAlignEnum.left :
-            (settings.align === textAlignEnum.right ? textAlignEnum.right :
-            (settings.align === textAlignEnum.justify ? textAlignEnum.justify :
-            (settings.align === textAlignEnum.start ? textAlignEnum.start : undefined))),
-
-        textColor: readiumCSSDefaults.textColor,
-
-        typeScale: readiumCSSDefaults.typeScale,
-
-        wordSpacing: settings.wordSpacing,
-
-        mathJax: settings.enableMathJax,
-
-        reduceMotion: readiumCSSDefaults.reduceMotion,
+    let mathJaxPath = "MathJax";
+    if (_PACKAGING === "1") {
+        mathJaxPath = path.normalize(path.join(__dirname, mathJaxPath));
+    } else {
+        mathJaxPath = "mathjax";
+        mathJaxPath = path.normalize(path.join(__dirname, _NODE_MODULE_RELATIVE_URL, mathJaxPath));
+    }
+    mathJaxPath = mathJaxPath.replace(/\\/g, "/");
+    debug("MathJax path:", mathJaxPath);
+    // https://expressjs.com/en/4x/api.html#express.static
+    const staticOptions = {
+        dotfiles: "ignore",
+        etag: true,
+        fallthrough: false,
+        immutable: true,
+        index: false,
+        maxAge: "1d",
+        redirect: false,
+        // extensions: ["css", "otf"],
+        setHeaders: (res: express.Response, _path: string, _stat: any) => {
+            //   res.set('x-timestamp', Date.now())
+            streamer.setResponseCORS(res);
+        },
     };
-    const jsonMsg: IEventPayload_R2_EVENT_READIUMCSS = { setCSS: cssJson };
-    return jsonMsg;
-}
+    const MATHJAX_URL_PATH = "math-jax";
+    streamer.expressUse("/" + MATHJAX_URL_PATH, express.static(mathJaxPath, staticOptions));
+    const transformer = (_publication: R2Publication, _link: Link, str: string): string => {
 
-setupReadiumCSS(streamer, rcssPath, computeReadiumCssJsonMessage);
+        const store = diMainGet("store");
+        const settings = store.getState().reader.config;
 
-let mathJaxPath = "MathJax";
-if (_PACKAGING === "1") {
-    mathJaxPath = path.normalize(path.join(__dirname, mathJaxPath));
-} else {
-    mathJaxPath = "mathjax";
-    mathJaxPath = path.normalize(path.join(__dirname, _NODE_MODULE_RELATIVE_URL, mathJaxPath));
-}
-mathJaxPath = mathJaxPath.replace(/\\/g, "/");
-debug("MathJax path:", mathJaxPath);
-// https://expressjs.com/en/4x/api.html#express.static
-const staticOptions = {
-    dotfiles: "ignore",
-    etag: true,
-    fallthrough: false,
-    immutable: true,
-    index: false,
-    maxAge: "1d",
-    redirect: false,
-    // extensions: ["css", "otf"],
-    setHeaders: (res: express.Response, _path: string, _stat: any) => {
-        //   res.set('x-timestamp', Date.now())
-        streamer.setResponseCORS(res);
-    },
-};
-const MATHJAX_URL_PATH = "math-jax";
-streamer.expressUse("/" + MATHJAX_URL_PATH, express.static(mathJaxPath, staticOptions));
-const transformer = (_publication: R2Publication, _link: Link, str: string): string => {
-
-    const store = diMainGet("store");
-    const settings = store.getState().reader.config;
-
-    if (settings.enableMathJax) {
-        const url = `${streamer.serverUrl()}/${MATHJAX_URL_PATH}/es5/tex-mml-chtml.js`;
-        const script = `
+        if (settings.enableMathJax) {
+            const url = `${streamer.serverUrl()}/${MATHJAX_URL_PATH}/es5/tex-mml-chtml.js`;
+            const script = `
         <script type="text/javascript">
 window.MathJax = {
     startup: {
@@ -168,9 +103,42 @@ window.MathJax = {
 };
         </script>
         <script type="text/javascript" async="async" src="${url}"> </script>`;
-        return str.replace(/<\/head>/, `${script}</head>`);
-    } else {
-        return str;
-    }
+            return str.replace(/<\/head>/, `${script}</head>`);
+        } else {
+            return str;
+        }
+    };
+    Transformers.instance().add(new TransformerHTML(transformer));
+
+    return streamer;
+}
+
+function startStreamer(streamer: Server) {
+    // Find a free port on your local machine start at 8000
+
+    return async () => {
+        const port = await portfinder.getPortPromise();
+        await streamer.start(port, true);
+
+        const streamerUrl = streamer.serverUrl();
+        debug("Streamer started on %s", streamerUrl);
+
+        return streamerUrl;
+    };
+}
+
+function stopStreamer(streamer: Server) {
+    return () => streamer.stop();
+}
+
+const streamerServer = streamerFactory();
+const start = startStreamer(streamerServer);
+const stop = stopStreamer(streamerServer);
+
+export {
+    streamerServer,
+    startStreamer,
+    stopStreamer,
+    start,
+    stop,
 };
-Transformers.instance().add(new TransformerHTML(transformer));
